% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agg.R
\name{Agg}
\alias{Agg}
\title{Find 3D aggregates}
\usage{
Agg(imgs, channel, kern.neighbour = c(3, 3, 3), kern.smooth = c(3, 3, 3),
  layers = NULL, pwidth = NULL, zstep = NULL)
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from loadIMG or findIMG functions. A dataframe from the quantify function}

\item{channel}{Name of the channel to find aggregates in. Should be in the names of the array files}

\item{kern.neighbour}{Numeric vector indicating range of neighbouring pixels to aggregate in the x,y,z directions. Has to be odd intergers. c(1,1,1) means no aggregating.}

\item{kern.smooth}{Numeric vector indicating range of median smoothing in the x,y,z directions. Has to be odd intergers. c(1,1,1) means no smoothing.}

\item{layers}{Optional. Should the function only look in a subset of layers. A list with lists of layers to use for each image. Can also be the output from ELayers}

\item{pwidth}{Optional. Width of pixels in microns to calculate aggregate size in microns instead of pixels}

\item{zstep}{Optional. z-step in microns to calculate aggregate size in microns instead of pixels}
}
\value{
A list with two parts. First part is a dataframe with ID and size of aggregates and name of image, second part is a list of the arrays in which pixels are NA if empty or given a number indicating the aggregate ID
}
\description{
Function to group adjacent pixels in aggregates
}
\keyword{aggregate}
\keyword{array}
\keyword{image}
