% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{center_of_mass}
\alias{center_of_mass}
\title{Find Center of Mass in images}
\usage{
center_of_mass(imgs, channels)
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from \code{loadIMG} or \code{findIMG} functions.}

\item{channels}{Character vector with name(s) of channels. Channel names should be in the names of the array files}
}
\value{
A dataframe with number of pixels for each layer
}
\description{
Find Center of Mass in images
}
\keyword{array}
\keyword{image}
\keyword{quantify}
