% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumps_dist.R
\name{clumps_dist}
\alias{clumps_dist}
\title{Find distances between 3D aggregates/clumps}
\usage{
clumps_dist(clumps.x, clumps.y = clumps.x, pwidth = 1, zstep = 1, cores = 1)
}
\arguments{
\item{clumps.x}{Data.frame output from the \code{clumps} function (...$Aggregates)}

\item{clumps.y}{Optional. Data.frame output from the \code{clumps} function (...$Aggregates)}

\item{pwidth}{Width of pixels in microns}

\item{zstep}{z-step in microns}

\item{cores}{Number of cores to use for parallel computing}
}
\value{
A dataframe with the distances
}
\description{
With two outputs from the \code{clumps} function compute the distances between pairs of aggregates. 
Computes the distances between aggregates in \code{clumps.x} with those in \code{clumps.y}.
If you want distances between aggregates from one output, just input the data.frame in \code{clumps.x}.
Provide pwidth and zstep to get the distance in microns. Alternatively it's the pixel distance
}
\keyword{aggregate}
\keyword{array}
\keyword{image}
