% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumps_plot.R
\name{clumps_plot}
\alias{clumps_plot}
\title{Plot 3D aggregates}
\usage{
clumps_plot(clumps.out, replica = 1, col = "agg", pwidth = NULL,
  zstep = NULL, center = NULL, radius = NULL, thresh = 0,
  thresh.m = FALSE)
}
\arguments{
\item{clumps.out}{Output from \code{clumps}}

\item{replica}{Image replica to plot}

\item{col}{Colour by aggregate ID ("agg") or by distance to a point ("dist")}

\item{pwidth}{Pixel width in microns}

\item{zstep}{Microns between the layers}

\item{center}{3D coordinates (c(x,y,z)) of where to center circular outline}

\item{radius}{Radius of the circular outline, in microns if pwidth and zstep are provided, else in pixels}

\item{thresh}{Clumps smaller than this are not plotted (in pixels)}

\item{thresh.m}{If TRUE will apply threshold to sizes in microns instead of pixels}
}
\description{
Provide pwidth and zstep to get true aspect ratio. Provide center and radius to draw a circular outline
}
\keyword{aggregate}
\keyword{array}
\keyword{image}
