% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_agg.R
\name{co_agg}
\alias{co_agg}
\alias{co_agg.default}
\title{3D co-aggregation}
\usage{
co_agg(
  imgs,
  channels,
  size,
  npixel,
  R = 10,
  dstep = 1,
  pwidth,
  zstep,
  cores = 1,
  kern.smooth = NULL,
  layers = NULL,
  naming = NULL
)

co_agg.default(
  imgs,
  channels,
  size,
  npixel,
  dstep = 1,
  pwidth,
  zstep,
  cores = 1,
  kern.smooth = NULL,
  layers = NULL,
  naming = NULL
)
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from \code{loadIMG} or \code{findIMG} functions.}

\item{channels}{Character vector with names of the two channels to calculate co-aggregation for. Should be in the names of the array files}

\item{size}{The maximum distance (microns) to examine. Has to be a multiple of both pwidth and zstep. Beware, increasing size will increase runtime exponetially!}

\item{npixel}{Number of random pixels to examine. Increasing this will increase precision (and runtime in a linear fashion)}

\item{R}{Number of times to run the co-aggregation analysis}

\item{dstep}{The interval between examined distances (microns). Increasing this decreases resolution but speeds up function linearly. Defaults to 1}

\item{pwidth}{Width of pixels in microns}

\item{zstep}{z-step in microns}

\item{cores}{The number of cores to use. Defaults to 1}

\item{kern.smooth}{Optional. Numeric vector indicating range of median smoothing in the x,y,z directions. Has to be odd intergers. c(1,1,1) means no smoothing.}

\item{layers}{Optional. Should the function only look in a subset of layers. A list with lists of layers to use for each image. Can also be the output from \code{extract_layers}}

\item{naming}{Optional. Add metadata to the output dataframe by looking through names of array files. Should be a list of character vectors, each list element will be added as a variable. Example: naming=list(Time=c("T0","T1","T2")). The function inserts a variable called Time, and then looks through the names of the array files and inserts characters mathcing either T0, T1 or T2}
}
\value{
A dataframe with the co-aggregation values for each distance
}
\description{
Function to calculate the pairwise 3D co-aggregation between two channels
}
\keyword{array}
\keyword{co-aggregation}
\keyword{image}
