% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random.R
\name{create_random}
\alias{create_random}
\title{Create random image}
\usage{
create_random(path, overlap, side, h, probs)
}
\arguments{
\item{path}{The path of where to save the RDS files}

\item{overlap}{Logical. Are the channels allowed to overlap, such that they can be at the same pixels}

\item{side}{Pixels in x,y dimension, a numeric}

\item{h}{Height of image, i.e. number of layers}

\item{probs}{Numeric vector with an element for each channel, these elements gives the probability for each pixel of being assigned to each channel.}
}
\value{
Creates arrays as RDS files in the specified path, and outputs the paths for these files
}
\description{
Create array of random image
}
\keyword{array}
\keyword{image}
