% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossCorRandom.R
\name{CrossCorRandom}
\alias{CrossCorRandom}
\title{3D random cross-correlation}
\usage{
CrossCorRandom(overlap, size, npixel, dstep, side, h, pwidth, zstep, R,
  p1 = 1/3, p2 = 1/3, freec = 1)
}
\arguments{
\item{overlap}{Logical. Are the channels allowed to overlap, such that they can be at the same pixels}

\item{size}{The maximum distance (microns) to examine. Has to be a multiple of both pwidth and zstep. Beware, increasing size will increase runtime exponetially!}

\item{npixel}{Number of random pixels to examine. Increasing this will increase precision (and runtime in a linear fashion)}

\item{dstep}{The interval between examined distances (microns). Increasing this decreases resolution but speeds up function linearly. Defaults to 1}

\item{side}{Pixels in x,y dimension, a numeric}

\item{pwidth}{Width of pixels in microns}

\item{zstep}{z-step in microns}

\item{R}{Number of runs}

\item{p1}{For each pixel, the probability of being assigned to channel 1.}

\item{p2}{For each pixel, the probability of being assigned to channel 2.}

\item{freec}{The number of cores NOT to use.Defaults to 1}
}
\value{
A dataframe with the cross-correlation vaules for each distance
}
\description{
Function to calculate the pairwise 3D cross-correlation between two random channels
}
\keyword{array}
\keyword{cross-correlation}
\keyword{image}
