% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossCorX.R
\name{CrossCorX}
\alias{CrossCorX}
\title{3D cross-correlation}
\usage{
CrossCorX(..., R = NULL)
}
\arguments{
\item{...}{Arguments to the CrossCor function}

\item{R}{Number of times to run the cross-correlation analysis}
}
\value{
A dataframe with the cross-correlation vaules for each distance
}
\description{
Wrapper function to run CrossCor several times
}
\keyword{array}
\keyword{cross-correlation}
\keyword{image}
