% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELayers.R
\name{ELayers}
\alias{ELayers}
\title{Extract layers}
\usage{
ELayers(splitdf, part, ...)
}
\arguments{
\item{splitdf}{A dataframe from the LayerSplit function}

\item{part}{What part should be extracted? Example: part="Top"}

\item{...}{Arguments for subsetting the splitdf dataframe first}
}
\value{
A list with lists of the extracted layers, for use in the Agg and CrossCor function
}
\description{
Function to extract layers from the LayerSplit function for use in the CrossCor or Agg analysis
}
\keyword{array}
\keyword{image}
\keyword{split}
