% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layers.R
\name{extract_layers}
\alias{extract_layers}
\title{Extract layers}
\usage{
extract_layers(splitdf, part, ...)
}
\arguments{
\item{splitdf}{A dataframe from the \code{layer_split} function}

\item{part}{What part should be extracted? Example: part="Top"}

\item{...}{Arguments for subsetting the splitdf dataframe first}
}
\value{
A list with lists of the extracted layers, for use in the \code{co_agg}, \code{occupancy} or \code{clumps} functions
}
\description{
Function to extract layers from the \code{layer_split} function for use in the \code{co_agg}, \code{occupancy} or \code{clumps} analysis
}
\keyword{array}
\keyword{image}
\keyword{split}
