% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_split.R
\name{layer_split}
\alias{layer_split}
\title{Split quantification in Top, Middle and Bottom}
\usage{
layer_split(qua, side.x, side.y, pt, add.t, add.b, channel = NULL,
  trim = FALSE, layer.start = NULL)
}
\arguments{
\item{qua}{A dataframe from the \code{quant} function}

\item{side.x}{x dimension of image in pixels}

\item{side.y}{y dimension of image in pixels}

\item{pt}{The Top is defined from layer 1 (or max if layer.start="Bottom") until, but not including, the layer where the image is filled by the proportion pt (0.9 = 90 percent). If pt is set higher than found in any layer in the image then add.b specifies the Bottom and the rest is the Top. Note that if pt, add.t and add.b is set such that there is overlap between Bottom, Middle and Top, then Top will override the others}

\item{add.t}{How many layers to add to the top, after pt calculation (e.g. pt=0 and add.t=10 will specify the top as the top 10 layers)}

\item{add.b}{How many layers are specified as the bottom}

\item{channel}{Name of channel to focus on. If NULL will aggregate all channels before splitting}

\item{trim}{Should the bottom be trimmed? Specifies the bottom as the first layer (from the bottom) with max counts. Defaults to FALSE}

\item{layer.start}{Are first layers the "Top" or the "Bottom" of the specimen?}
}
\value{
The original qua dataframe including an Exp variable, with a unique name for each image (based on the naming given in the \code{quant} function). And a Split varaible denoting the split
}
\description{
Function to split quantification in Top, Middle and Bottom layer by proportion occupied
}
\keyword{array}
\keyword{image}
\keyword{split}
