% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerStd.R
\name{LayerStd}
\alias{LayerStd}
\title{Trim and standardize quantification}
\usage{
LayerStd(qua, trim = FALSE, layer.start = NULL)
}
\arguments{
\item{qua}{A dataframe from the quantify function}

\item{trim}{Should the bottom be trimmed? Specifies the bottom as the first layer (from the bottom) with max counts. Defaults to FALSE}

\item{layer.start}{Are first layers the "Top" or the "Bottom" of the specimen? If equal to "Top", the layers will be reversed so that low numbers denote the bottom of the specimen}
}
\value{
The original qua dataframe including an Exp variable, with a unique name for each image (based on the naming given in the quantify function). And a NewLayer varaible denoting the layer from the bottom of the specimen after trimming
}
\description{
Function to trim and standardize quantification
}
\keyword{array}
\keyword{image}
\keyword{split}
