% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIMG.R
\name{loadIMG}
\alias{loadIMG}
\title{Load .tif images into R}
\usage{
loadIMG(path, channels, split = FALSE)
}
\arguments{
\item{path}{The path of either a folder with all .tif files to be loaded (if split=FALSE), or a folder where each image is in a separate subfolder with all z-stacks and channels in that folder (if split=TRUE)}

\item{channels}{Character vector with name(s) of channels. Channel names should be in the names of the .tif files}

\item{split}{Logical indicating if all z-stacks are in one .tif (FALSE) or in a folder with a .tif for each z-stack (TRUE)}
}
\value{
Creates arrays as RDS files in working directory, and outputs the paths for these files
}
\description{
Create an array from one or more .tif files. Looks for .tif in the working directory
}
\keyword{array}
\keyword{image}
\keyword{tif}
