% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphIMG.R
\name{morphIMG}
\alias{morphIMG}
\title{Apply mathematical morphology kernel on image and save as RDS}
\usage{
morphIMG(imgs, morph = NULL, kern = c(3, 3, 3), type = "box",
  cores = 1)
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from \code{loadIMG} or \code{findIMG} functions.}

\item{morph}{One of "erode", "dilate", "opening", or "closing"}

\item{kern}{Numeric vector indicating range of morpinh in the x,y,z directions. Has to be odd integers.}

\item{type}{Type of kernel for morphing "box" includes diagonals, "diamond" is without diagonals}

\item{cores}{Integer. Number of cores to use for parallel computing.}
}
\value{
Creates arrays as RDS files and outputs the paths for these files
}
\description{
Morph images as a preparation for subsequent analyses
}
\keyword{array}
\keyword{image}
