% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantify.R
\name{Quantify}
\alias{Quantify}
\title{Quantifies the pixels in images}
\usage{
Quantify(imgs, channels, naming = NULL)
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from loadIMG or findIMG functions.}

\item{channels}{Character vector with name(s) of channels. Channel names should be in the names of the array files}

\item{naming}{Optional. Add metadata to the output dataframe by looking through names of array files. Should be a list of character vectors, each list element will be added as a variable. Example: naming=list(Time=c("T0","T1","T2")). The function inserts a variable called Time, and then looks through the names of the array files and inserts characters mathcing either T0, T1 or T2}
}
\value{
Creates arrays as RDS files in working directory, and outputs the paths for these files
}
\description{
Quantifies the number of pixels for each channel in each layer (z-stack)
}
\keyword{array}
\keyword{image}
\keyword{quantify}
