% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshIMG.R
\name{threshIMG}
\alias{threshIMG}
\title{Apply thresholding on image and save as RDS}
\usage{
threshIMG(imgs, method = "Manual", cores = 1, breaks = 100, opt = 0.1,
  thresh = rep(0, length(imgs)), BEM.opt = list(start = list(a = -1e+06, b =
  0.1, c = 1e+06), control = nls.lm.control(maxiter = 1000)))
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from \code{loadIMG} or \code{findIMG} functions.}

\item{method}{Either Manual, Otsu, or BEM}

\item{cores}{Integer. Number of cores to use for parallel computing.}

\item{breaks}{Integer. Number of breaks for Otsu's threshold. Increase for increased precision}

\item{opt}{Numeric. Between 0 and 1. Parameter for BEM method. Increase to decrease threshold}

\item{thresh}{Numeric. Vector of thresholds for manual thresholding, one for each image.}

\item{BEM.opt}{List. A list of options for model fitting for the BEM method. Passed to the nlsLM function.}
}
\value{
Prints thresholds. saves thresholded RDS files and outputs the path for the thresholded images.
}
\description{
Thresholding
}
\keyword{array}
\keyword{image}
