% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiffToArray.R
\name{tiffToArray}
\alias{tiffToArray}
\title{Make Array from .tif file(s)}
\usage{
tiffToArray(channels, split)
}
\arguments{
\item{channels}{Character vector with name(s) of channels. Channel names should be in the names of the .tif files}

\item{split}{Logical indicating if all z-stacks are in one .tif (FALSE) or in a folder with a .tif for each z-stack (TRUE)}
}
\value{
Arrays as RDS files in working directory
}
\description{
Create an array from one or more .tif files. Looks for .tif in the working directory
}
\keyword{array}
\keyword{tif}
