% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_splits.R
\name{xy_splits}
\alias{xy_splits}
\title{Analyses for each xy position}
\usage{
xy_splits(imgs, channels, do, upper.part = 0.5, layer.start = NULL,
  cores = 1)
}
\arguments{
\item{imgs}{The paths of array files; i.e. output from \code{loadIMG} or \code{findIMG} functions.}

\item{channels}{Character vector with name(s) of channels. Channel names should be in the names of the array files}

\item{do}{What to do? \itemize{
\item "max" and "min" find the max or min layer where the first 1 is found.
\item "sum" sums the pixels (e.g. for biomass xy-distribution). 
\item "which" returns z-positions of pixels for each xy dimension. 
\item "section" first sections the image in an upper/lower part and then sums for each part. Sectioning is done for each xy-position and therefore takes a variable biomass distribution across the image into account.
\item Alternatively a user-defined function that will be applied for each xy-position. E.g. function(x) median(which(x > 0)) will return the median position of each channel at each xy-position.
}}

\item{upper.part}{Only if do="section". Either the fraction to assign to the upper section (between 0 and 1), or a fixed number of layers that are assigned to the upper part (an integer, 1L, 2L, 3L etc.. Remember the L!). The upper fraction is inclusive (i.e. x >= upper.part)}

\item{layer.start}{Only if do="section". Are first layers the "Top" or the "Bottom" of the specimen?}

\item{cores}{Cores to use for parallel computing. Defaults to 1, i.e. not parallel.}
}
\value{
A recursive list. First layer are the images, within those are the channels. Each element is a matrix with a value for each xy position. If do="section" an extra layer is added denoting the upper and lower section.
}
\description{
Splits the image in each xy position and then run an analysis
}
\details{
For sectioning, NA is returned if a channel is absent at a specific xy-position, but a zero is returned if the channel is present at that xy-position but absent in the section.
}
\keyword{array}
\keyword{image}
