% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIMG.R
\name{findIMG}
\alias{findIMG}
\title{Find arrays created with loadIMG or tiffToArray}
\usage{
findIMG(path)
}
\arguments{
\item{path}{The path of the _Array.R files}
}
\value{
The paths for the array files
}
\description{
Find arrays already created by the loadIMG or tiffToArray functions
}
\keyword{array}
\keyword{image}
\keyword{tif}
