% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIMG.R
\name{findIMG}
\alias{findIMG}
\title{Find arrays created with \code{loadIMG} or \code{tiff_to_array}}
\usage{
findIMG(path, channels = NULL)
}
\arguments{
\item{path}{The path of the _Array.R files}

\item{channels}{Optional. Channel names to look for in the image files}
}
\value{
The paths for the array files
}
\description{
Find arrays already created by the \code{loadIMG} or \code{tiff_to_array} functions
}
\keyword{array}
\keyword{image}
\keyword{tif}
