% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_channels.R
\name{merge_channels}
\alias{merge_channels}
\title{Create RDS for a combination of channels}
\usage{
merge_channels(path, channels, method)
}
\arguments{
\item{path}{The path of the _Array.R files}

\item{channels}{Character vector with names of the channels. Channel names should be in the names of the .tif files}

\item{method}{How to merge. Either "union" (presence is coded if at least one channel is present), "intersect" (presence is coded only if all channels are present), or "subtract" (presence is coded as those in the first channel but not in the second channel).}
}
\value{
Creates arrays as RDS files in the specified path, and outputs the paths for these files
}
\description{
Create RDS for a combination of channels
}
\keyword{array}
\keyword{image}
